youtube-blog/requirements.txt
fastapi
uvicorn
requests
openai
youtube-transcript-api
python-docx
reportlab